Refactor DTOs locations
