+++ interview-webapp/src/main/java/com/example/interviewWebapp/Dto/QuestionDTO/QuestionResponseDTO.java
@@ -1,4 +1,4 @@
-package com.example.interviewWebapp.Dto;
+package com.example.interviewWebapp.Dto.QuestionDTO;
 
 import java.util.List;
 
@@ -10,6 +10,7 @@ public class QuestionResponseDTO {
     private String category;
     private String type;
     private List<String> options;
+    private String correctAnswer;
 
     public String getId() {
         return id;
@@ -65,6 +66,14 @@ public class QuestionResponseDTO {
 
     public void setOptions(List<String> options) {
         this.options = options;
+    }
+
+    public String getCorrectAnswer() {
+        return correctAnswer;
+    }
+
+    public void setCorrectAnswer(String correctAnswer) {
+        this.correctAnswer = correctAnswer;
     }
 }
