+++ interview-webapp/src/main/java/com/example/interviewWebapp/Mapper/InterviewMapper.java
@@ -1,8 +1,8 @@
 package com.example.interviewWebapp.Mapper;
 
 
-import com.example.interviewWebapp.Dto.InterviewResponseDTO;
-import com.example.interviewWebapp.Dto.StartInterviewRequestDTO;
+import com.example.interviewWebapp.Dto.InterviewsDTO.InterviewResponseDTO;
+import com.example.interviewWebapp.Dto.InterviewsDTO.StartInterviewRequestDTO;
 import com.example.interviewWebapp.Entity.Interviews;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Component;
