+++ interview-webapp/src/main/java/com/example/interviewWebapp/Mapper/ResponseMapper.java
@@ -1,7 +1,7 @@
 package com.example.interviewWebapp.Mapper;
 
-import com.example.interviewWebapp.Dto.GetResponsesDTO;
-import com.example.interviewWebapp.Dto.SubmitResponseRequestDTO;
+import com.example.interviewWebapp.Dto.ResponsesDTO.GetResponsesDTO;
+import com.example.interviewWebapp.Dto.ResponsesDTO.SubmitResponseRequestDTO;
 import com.example.interviewWebapp.Entity.Responses;
 import org.bson.types.ObjectId;
 import org.modelmapper.ModelMapper;
