+++ interview-webapp/src/main/java/com/example/interviewWebapp/Service/ResponseService.java
@@ -1,6 +1,8 @@
 package com.example.interviewWebapp.Service;
 
-import com.example.interviewWebapp.Dto.*;
+import com.example.interviewWebapp.Dto.ResponsesDTO.GetResponsesDTO;
+import com.example.interviewWebapp.Dto.ResponsesDTO.SubmitMultipleResponsesRequestDTO;
+import com.example.interviewWebapp.Dto.ResponsesDTO.SubmitResponseRequestDTO;
 import com.example.interviewWebapp.Entity.Interviews;
 import com.example.interviewWebapp.Entity.Questions;
 import com.example.interviewWebapp.Entity.Responses;
@@ -9,14 +11,10 @@ import com.example.interviewWebapp.Repository.InterviewRepo;
 import com.example.interviewWebapp.Repository.QuestionRepo;
 import com.example.interviewWebapp.Repository.ResponseRepo;
 import org.bson.types.ObjectId;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
-import java.util.stream.Collectors;
 
 @Service
 public class ResponseService {
